Extracted repository methods from RoomService. Removed methods for old functionality that is no longer needed
