+++ hotel-booking-app-service/src/main/java/com/application/hotelbooking/transformers/RoomTransformer.java
@@ -1,6 +1,7 @@
 package com.application.hotelbooking.transformers;
 
 import com.application.hotelbooking.domain.*;
+import com.application.hotelbooking.dto.RoomCreationServiceDTO;
 import com.application.hotelbooking.dto.RoomDTO;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -31,5 +32,10 @@ public class RoomTransformer {
 
     public RoomDTO transformToRoomDTO(Room room){
         return modelMapper.map(room, RoomDTO.class);
+    }
+
+    public Room transformToRoom(RoomCreationServiceDTO roomCreationServiceDTO){
+        modelMapper.typeMap(RoomCreationServiceDTO.class, Room.class).addMappings(modelMapper -> modelMapper.skip(Room::setId));
+        return modelMapper.map(roomCreationServiceDTO, Room.class);
     }
 }
